+++ src/main/java/vip/gadfly/chakkispring/service/impl/ClassManageServiceImpl.java
@@ -39,7 +39,7 @@ public class ClassManageServiceImpl extends ServiceImpl<ClassMapper, ClassDO> im
     }
 
     @Override
-    public IPage<ClassDO> getClassPage(long page, long count) {
+    public IPage<ClassDO> getClassPage(Integer page, Integer count) {
         Page<ClassDO> pager = new Page<>(page, count);
         return this.baseMapper.selectPage(pager, null);
     }
